refactor: Update favicon and page title in index.html
Update the favicon and page title in the index.html file to use the new favicon.jpg and set the title to "Kaiju". This change improves the branding and visual identity of the application.

this is copilot gen message. i have made a lot of changes but to lazy to write them down here look yourself.